+++ src/Hydrator.php
@@ -145,7 +145,7 @@ final class Hydrator implements HydratorInterface
                 );
                 if ($result->isResolved()) {
                     $this
-                        ->preparePropertyToSetValue($reflectionClass, $property, $propertyName)
+                        ->preparePropertyToSetValue($reflectionClass, $property)
                         ->setValue($object, $result->getValue());
                 }
             }
@@ -155,14 +155,15 @@ final class Hydrator implements HydratorInterface
     private function preparePropertyToSetValue(
         ReflectionClass $class,
         ReflectionProperty $property,
-        string $propertyName,
     ): ReflectionProperty {
-        // It needs for PHP 8.3 and earlier versions.
-        if ($property->isReadOnly()) {
+        if (
+            (PHP_VERSION_ID < 80400 && $property->isReadOnly())
+            || (PHP_VERSION_ID >= 80400 && $property->isPrivateSet())
+        ) {
             $declaringClass = $property->getDeclaringClass();
-            if ($declaringClass !== $class) {
-                return $declaringClass->getProperty($propertyName);
-            }
+            return $declaringClass->getName() === $class->getName()
+                ? $property
+                : $declaringClass->getProperty($property->getName());
         }
 
         return $property;