+++ tests/TestEnvironments/Php84/Hydrator/PublicPrivateSetProperty/Figure.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Hydrator\Tests\TestEnvironments\Php84\Hydrator\PublicPrivateSetProperty;
+
+abstract class Figure
+{
+    public private(set) ?int $radius = null;
+}